+++ docs/guides/changelog.md
@@ -7,6 +7,10 @@ description: |-
 
 # Changelog
 
+## 0.2.4 (unreleased)
+
+- Add a mutex to prevent multiple concurrent operations (such as Create, Read, Update, or Delete) on the same resource instance or API endpoint
+
 ## 0.2.3
 
 - BREAKING CHANGE: Rename resource `catalystcenter_fabric_port_assignment` to `catalystcenter_fabric_port_assignments` and fix issue with adding/removing ports to/from `catalystcenter_fabric_port_assignments` resource